+++ src/use-lunatic/commons/index.ts
@@ -1,6 +1,10 @@
-export { default as checkLoops } from './check-loops';
-export { default as createMapPages } from './create-map-pages';
+export { default as checkLoops } from './pagination/check-loops';
+export { default as createMapPages } from './pagination/create-map-pages';
 export { default as executeConditionFilter } from './execute-condition-filter';
 export { default as getCompatibleVTLExpression } from './get-compatible-vtl-expression';
-export { default as isFirstLastPage } from './is-First-last-page';
-export { getNewReachedPage, getPageTag, isNewReachedPage } from './page-tag';
+export { default as isFirstLastPage } from './pagination/is-First-last-page';
+export {
+	getNewReachedPage,
+	getPageTag,
+	isNewReachedPage,
+} from './pagination/page-tag';
