+++ src/use-lunatic/commons/pagination/is-First-last-page.ts
@@ -1,4 +1,4 @@
-import type { LunaticReducerState } from '../type';
+import type { LunaticReducerState } from '../../type';
 
 function isFirstLastPage(pager: LunaticReducerState['pager']) {
 	const { page, maxPage } = pager;
