+++ src/use-lunatic/commons/pagination/is-paginated-loop.ts
@@ -1,4 +1,4 @@
-import type { LunaticComponentDefinition } from '../type';
+import type { LunaticComponentDefinition } from '../../type';
 
 function isPaginatedLoop(
 	component: LunaticComponentDefinition
