+++ src/use-lunatic/commons/pagination/is-roundabout.ts
@@ -1,4 +1,4 @@
-import type { LunaticComponentDefinition } from '../type';
+import type { LunaticComponentDefinition } from '../../type';
 
 export default function isRoundabout(
 	component: LunaticComponentDefinition
