+++ src/use-lunatic/commons/pagination/page.ts
@@ -1,6 +1,9 @@
-import type { LunaticComponentDefinition, LunaticReducerState } from '../type';
-import { getComponentsFromState } from './get-components-from-state';
-import executeConditionFilter from './execute-condition-filter';
+import type {
+	LunaticComponentDefinition,
+	LunaticReducerState,
+} from '../../type';
+import { getComponentsFromState } from '../get-components-from-state';
+import executeConditionFilter from '../execute-condition-filter';
 
 export function getPageId({
 	subPage,
